+++ src/components/LoansByCategory/QuickFilters/QuickFilters.vue
@@ -14,7 +14,9 @@
 				Quick filters
 			</h3>
 			<span v-show="filtersLoaded" class="tw-ml-2 tw-text-small">Showing {{ totalLoans }} loans</span>
-			<span v-show="filtersLoaded" class="tw-ml-2 tw-text-small">Reset filters</span>
+			<button v-show="filtersLoaded" class="tw-ml-2 tw-text-small tw-text-action" @click="resetFilters">
+				Reset filters
+			</button>
 		</div>
 		<div class="tw-flex tw-flex-col lg:tw-flex-row tw-gap-2 tw-w-full tw-pr-2 lg:tw-pr-0">
 			<div class="tw-flex tw-flex-col tw-order-2 lg:tw-order-1">
@@ -40,7 +42,26 @@
 				:regions="filterOptions.location"
 				:total-loans="totalLoans"
 				:filters-loaded="filtersLoaded"
+				:update-location="updateLocation"
 			/>
+
+			<div class="tw-flex tw-flex-col tw-order-3">
+				<label
+					class="tw-text-h4"
+					for="gender"
+				>
+					Sort By
+				</label>
+				<kv-select :disabled="!filtersLoaded" v-model="sortBy" id="sortBy" style="min-width: 180px;">
+					<option
+						v-for="sortType in filterOptions.sorting"
+						:key="sortType.key"
+						:value="sortType.key"
+					>
+						{{ sortType.title }}
+					</option>
+				</kv-select>
+			</div>
 		</div>
 	</div>
 </template>
@@ -64,6 +85,10 @@ export default {
 		filtersLoaded: {
 			type: Boolean,
 			default: false
+		},
+		updateFilters: {
+			type: Function,
+			required: true
 		}
 	},
 	components: {
@@ -73,8 +98,28 @@ export default {
 	data() {
 		return {
 			selectedGender: '',
+			sortBy: 'personalized',
 			showBadge: false
 		};
+	},
+	watch: {
+		selectedGender(gender) {
+			this.updateFilters({ gender });
+		},
+		sortBy(sortBy) {
+			this.updateFilters({ sortBy });
+		}
+	},
+	methods: {
+		updateLocation(location) {
+			this.updateFilters({ country: location });
+		},
+		resetFilters() {
+			this.$emit('reset-filters');
+			this.selectedGender = '';
+			this.sortBy = 'personalized';
+			this.updateLocation([]);
+		}
 	},
 	mounted() {
 		const badgeCookie = this.cookieStore.get('quick_filter_new_badge') === 'true' || false;